# cmd/api/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "RTSP → HLS 변환 파이프라인 관리 및 상태 조회 API",
        "title": "GStreamer Media Server API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/pipeline": {
            "put": {
                "description": "RTSP 정보를 받아 새 파이프라인을 생성하고 DB에 저장합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline"
                ],
                "summary": "Create a pipeline (PUT)",
                "parameters": [
                    {
                        "description": "새 파이프라인 RTSP 정보",
                        "name": "rtsp",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/address.RTSPInformation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "파이프라인 생성 성공",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.responseSchema"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 본문",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.responseSchema"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "409": {
                        "description": "파이프라인이 이미 존재함",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.responseSchema"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "파이프라인 생성 또는 DB 저장 실패",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.responseSchema"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "기존 파이프라인을 새 설정으로 업데이트하고 DB에 반영합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline"
                ],
                "summary": "Update a pipeline (POST)",
                "parameters": [
                    {
                        "description": "업데이트할 RTSP 정보",
                        "name": "rtsp",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/address.RTSPInformation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "파이프라인 업데이트 성공",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.responseSchema"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/address.RTSPInformation"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 본문",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.responseSchema"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "파이프라인을 찾을 수 없음",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.responseSchema"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "파이프라인 업데이트 또는 DB 반영 실패",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.responseSchema"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/pipeline/active/all": {
            "get": {
                "description": "동작중인 모든 활성 파이프라인 정보를 조회하여 GetPipeLineResponse 형태로 반환합니다",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline"
                ],
                "summary": "Get all active pipelines",
                "responses": {
                    "200": {
                        "description": "활성 파이프라인 조회 성공",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.responseSchema"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/main.GetPipeLineResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "파이프라인 조회 실패",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.responseSchema"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/pipeline/all": {
            "get": {
                "description": "모든 파이프라인 설정 정보를 DB에서 조회하여 RTSPInformation 형태로 반환합니다",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline"
                ],
                "summary": "Get all pipelines",
                "responses": {
                    "200": {
                        "description": "모든 파이프라인 조회 성공",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.responseSchema"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/address.RTSPInformation"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "파이프라인 조회 실패",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.responseSchema"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "모든 파이프라인을 정리하고 DB에서도 모두 삭제합니다",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline"
                ],
                "summary": "Delete all pipelines",
                "responses": {
                    "200": {
                        "description": "모든 파이프라인 삭제 성공",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.responseSchema"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "파이프라인 삭제 실패",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.responseSchema"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/pipeline/{streamName}": {
            "get": {
                "description": "streamName에 해당하는 파이프라인 정보를 DB에서 조회하여 반환합니다",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline"
                ],
                "summary": "Get a pipeline by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "파이프라인 이름",
                        "name": "streamName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "파이프라인 조회 성공",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.responseSchema"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/main.GetPipeLineResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "파이프라인을 찾을 수 없음",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.responseSchema"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "streamName 파이프라인을 정지하고 삭제한 후 DB에서도 하드 삭제합니다",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline"
                ],
                "summary": "Delete a pipeline",
                "parameters": [
                    {
                        "type": "string",
                        "description": "삭제할 파이프라인 이름",
                        "name": "streamName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "파이프라인 삭제 성공",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.responseSchema"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "파이프라인 삭제 실패",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.responseSchema"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/pipeline/{streamName}/start": {
            "post": {
                "description": "streamName 파이프라인을 실행 상태로 변경하고 DB에 상태를 업데이트합니다",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline"
                ],
                "summary": "Start a pipeline",
                "parameters": [
                    {
                        "type": "string",
                        "description": "시작할 파이프라인 이름",
                        "name": "streamName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "파이프라인 시작 성공",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.responseSchema"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "파이프라인 시작 실패",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.responseSchema"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/pipeline/{streamName}/status": {
            "get": {
                "description": "streamName 파이프라인의 실시간 상태 정보를 반환합니다 (TODO: 구현 예정)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline"
                ],
                "summary": "Get pipeline status by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "파이프라인 이름",
                        "name": "streamName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "파이프라인 상태 조회",
                        "schema": {
                            "$ref": "#/definitions/main.responseSchema"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "address.RTSPInformation": {
            "type": "object",
            "properties": {
                "bitrate": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "max_files": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "playlist_length": {
                    "type": "integer"
                },
                "target_duration": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                },
                "use_gpu": {
                    "type": "boolean"
                }
            }
        },
        "main.GetPipeLineResponse": {
            "type": "object",
            "properties": {
                "bitrate": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "hls_url": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "max_files": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "playlist_length": {
                    "type": "integer"
                },
                "rtsp_url": {
                    "type": "string"
                },
                "target_duration": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "use_gpu": {
                    "type": "boolean"
                }
            }
        },
        "main.responseSchema": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Additional Data"
                },
                "message": {
                    "description": "success, fail",
                    "type": "string"
                },
                "status": {
                    "description": "HTTP Status Code",
                    "type": "integer"
                }
            }
        }
    }
}